requirements.txt
langchain
pypdf
streamlit
faiss-cpu
boto3
python-dotenv
langchain-community
streamlit_lottie